+++ relayer-cli/src/commands.rs
@@ -13,7 +13,8 @@ use crate::config::Config;
 
 use self::{
     create::CreateCmds, keys::KeysCmd, light::LightCmd, listen::ListenCmd, query::QueryCmd,
-    start::StartCmd, start_multi::StartMultiCmd, tx::TxCmd, version::VersionCmd,
+    start::StartCmd, start_multi::StartMultiCmd, tx::TxCmd, update::UpdateCmds,
+    version::VersionCmd,
 };
 
 mod cli_utils;
@@ -26,6 +27,7 @@ mod query;
 mod start;
 mod start_multi;
 mod tx;
+mod update;
 mod version;
 
 /// Default configuration file path
@@ -54,10 +56,14 @@ pub enum CliCmd {
     Light(LightCmd),
 
     /// The `create` subcommand
-    #[options(
-        help = "Create objects on chains (by performing connection or channel open handshakes)"
-    )]
+    #[options(help = "Create objects (client, connection, or channel) on chains")]
     Create(CreateCmds),
+
+    /// The `update` subcommand
+    #[options(
+        help = "Update objects on chains. Currently this sub-commands serves only to update clients"
+    )]
+    Update(UpdateCmds),
 
     /// The `start` subcommand
     #[options(help = "Start the relayer")]